Add files via upload
A Python programming language question paper has been designed to assist college students in preparing for their semester exams. The paper focuses on the basics of Python and is intended for students pursuing a course that includes Python as a subject. Its purpose is to help students revise and practice basic concepts to perform well in exams.